Modifications
include Exploratory Data Analysis (EDA) for some
intervals of number of cases
include EDA for correlation of weather factors with
number of cases
Specify in the objectives
a. the climate data that will be included
b. Metrics
c. comparison with traditional machine learning
models
d. how to assess the usability and effectiveness of the
dashboard through user feedback from health
professionals and policymakers